Generate a list of fictional country names

Generate a list of fictional country names

Usage

generate_states(n = 10, prefixed = 0.15, suffixed = 0.15)

Arguments

n

Integer number of country names to generate from a library of fictional country names.

prefixed

Proportion of country names with a prefix by default 0.15.

suffixed

Proportion of country names with a suffix by default 0.15.

Value

String vector of fictional country names

Examples

  generate_states(12)
#>  [1] "Westenam"               "Ugraria"                "Boroland United"       
#>  [4] "Osmayya"                "Empire of Aramin Union" "Maliwar"               
#>  [7] "Jawar"                  "Astan"                  "Telti"                 
#> [10] "Central Etror"          "Rhonda"                 "Lanfal"