Extract V-Dem and V-Party data

Functions to import the V-Dem and the V-Party datasets from the {vdemdata} package in a many packages universe consistent format.

import_vdem()

import_vparty()

Value

A dataframe of the[vdem] dataset in a many packages universe-consistent format.

A dataframe of the[vparty] dataset consistent with the many packages universe.

Details

import_vdem() imports VDem 11.1 dataset and formats it to be consistent with the many packages universe.

import_vparty() imports the V-Party dataset and formats it to be consistent with the many packages universe.

Examples

# \donttest{
import_vdem()
#> # A tibble: 27,380 × 4,171
#>    VDem_ID Abbrv Label  Country Beg   End   Year  Date  count…¹  year historic…²
#>      <dbl> <chr> <chr>  <chr>   <mda> <mda> <mda> <mda> <chr>   <dbl> <date>    
#>  1       3 MEX   Vicer… Mexico  1789  1820  1789  1789… Mexico   1789 1789-12-31
#>  2       3 MEX   Vicer… Mexico  1789  1820  1790  1790… Mexico   1790 1790-12-31
#>  3       3 MEX   Vicer… Mexico  1789  1820  1791  1791… Mexico   1791 1791-12-31
#>  4       3 MEX   Vicer… Mexico  1789  1820  1792  1792… Mexico   1792 1792-12-31
#>  5       3 MEX   Vicer… Mexico  1789  1820  1793  1793… Mexico   1793 1793-12-31
#>  6       3 MEX   Vicer… Mexico  1789  1820  1794  1794… Mexico   1794 1794-12-31
#>  7       3 MEX   Vicer… Mexico  1789  1820  1795  1795… Mexico   1795 1795-12-31
#>  8       3 MEX   Vicer… Mexico  1789  1820  1796  1796… Mexico   1796 1796-12-31
#>  9       3 MEX   Vicer… Mexico  1789  1820  1797  1797… Mexico   1797 1797-12-31
#> 10       3 MEX   Vicer… Mexico  1789  1820  1798  1798… Mexico   1798 1798-12-31
#> # … with 27,370 more rows, 4,160 more variables: histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, gapstart1 <dbl>, gapstart2 <dbl>,
#> #   gapstart3 <dbl>, gapend1 <dbl>, gapend2 <dbl>, gapend3 <dbl>,
#> #   gap_index <dbl>, COWcode <dbl>, v2x_polyarchy <dbl>,
#> #   v2x_polyarchy_codelow <dbl>, v2x_polyarchy_codehigh <dbl>,
#> #   v2x_polyarchy_sd <dbl>, v2x_libdem <dbl>, v2x_libdem_codelow <dbl>,
#> #   v2x_libdem_codehigh <dbl>, v2x_libdem_sd <dbl>, v2x_partipdem <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
# }
# \donttest{
import_vparty()
#> # A tibble: 11,898 × 389
#> # Groups:   VParty_ID [3,467]
#>    VPart…¹ Label Abbrv Country Count…² Count…³ Beg   End   Year  v2pae…⁴ party…⁵
#>      <dbl> <chr> <chr> <chr>   <chr>     <dbl> <mda> <mda> <mda> <chr>   <chr>  
#>  1    4908 inde… indep Afghan… Kingdo…      36 1965  2010  1965  indepe… 1970-2…
#>  2    4908 inde… indep Afghan… Kingdo…      36 1965  2010  1969  indepe… 1970-2…
#>  3    4908 inde… indep Afghan… Islami…      36 1965  2010  2005  indepe… 1970-2…
#>  4    4908 inde… indep Afghan… Islami…      36 1965  2010  2010  indepe… 1970-2…
#>  5    5184 Peop… HDK   Afghan… Democr…      36 1988  1988  1988  People… NA     
#>  6     190 Demo… PDS   Albania Republ…      12 1991  2017  1991  Democr… NA     
#>  7     190 Demo… PDS   Albania Republ…      12 1991  2017  1992  Democr… NA     
#>  8     190 Demo… PDS   Albania Republ…      12 1991  2017  1996  Democr… NA     
#>  9     190 Demo… PDS   Albania Republ…      12 1991  2017  1997  Democr… NA     
#> 10     190 Demo… PDS   Albania Republ…      12 1991  2017  2001  Democr… NA     
#> # … with 11,888 more rows, 378 more variables: country_name <chr>,
#> #   histname <chr>, country_text_id <chr>, year <dbl>, historical_date <date>,
#> #   codingstart <dbl>, gapstart <dbl>, gapend <dbl>, codingend <dbl>,
#> #   project <dbl>, COWcode <int>, gap_index <dbl>, v2xpa_antiplural <dbl>,
#> #   v2xpa_antiplural_codelow <dbl>, v2xpa_antiplural_codehigh <dbl>,
#> #   v2xpa_popul <dbl>, v2xpa_popul_codelow <dbl>, v2xpa_popul_codehigh <dbl>,
#> #   v2paseatshare <dbl>, v2panumbseat <dbl>, v2patotalseat <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
# }