regimes database documentation

regimes database documentation

regimes

Format

The regimes database is a list that contains the following 1 datasets: Polity5. For more information and references to each of the datasets used, please use the data_source() and data_contrast() functions.

Polity5:

A dataset with 17574 observations and the following 33 variables: Polity5_ID, year, Label, p5, cyear, flag, fragment, democ, autoc, polity, polity2, durable, xrreg, xrcomp, xropen, xconst, parreg, parcomp, exrec, exconst, polcomp, prior, eprec, interim, bprec, post, change, d5, sf, regtrans, Beg, End, speccat.