States datacube

The manystates::states datacube is a list containing 3 datasets: ISD, GW, and GGO. It is a work-in-progress, so please let us know if you have any comments or suggestions.

Usage

states

Format

ISD:

A dataset with 499 observations and 36 variables: stateID, StateName, Begin, End, StateNameAlt, Latitude, Longitude, StartType, EndType, cowID, cowNR, ISD_Category, Region, Start_Am, EStart_Am, Declare, DecDate, Population, ..., VioEnd, and VioEnd_Am.

HUGGO:

A dataset with observations and variables: .

GW:

A dataset with 216 observations and 7 variables: stateID, StateName, Begin, End, StateNameAlt, cowID, and cowNR.

For more information and references to each of the datasets used, please use the manydata::call_sources() and manydata::compare_dimensions() functions.

Source

Griffiths, Ryan D., and Charles R. Butcher. 'Introducing the international system(s) dataset (ISD), 1816-2011'. International Interactions 39.5 (2013), pp. 748-768.

Gleditsch, Kristian S., and Michael D. Ward. 'Interstate system membership: A revised list of the independent states since 1816'. International Interactions 25.4 (1999), pp. 393-413.

Hand-coded data by the GGO team

Details

#> $ISD
#> ---------------------------------------------------------
#> |   Variable   |  Class  |  Obs  |  Missing  |  Miss %  |
#> ---------------------------------------------------------
#> |stateID       |character|    499|          0|         0|
#> |StateName     |character|    499|          0|         0|
#> |Begin         |mdate    |    282|        217|     43.49|
#> |End           |mdate    |    499|          0|         0|
#> |StateNameAlt  |character|    210|        289|     57.92|
#> |Latitude      |character|    343|        156|     31.26|
#> |Longitude     |character|    343|        156|     31.26|
#> |StartType     |numeric  |    337|        162|     32.46|
#> |EndType       |numeric  |    296|        203|     40.68|
#> |cowID         |character|    499|          0|         0|
#> |cowNR         |numeric  |    499|          0|         0|
#> |ISD_Category  |numeric  |    497|          2|       0.4|
#> |Region        |numeric  |    499|          0|         0|
#> |Start_Am      |numeric  |    499|          0|         0|
#> |EStart_Am     |numeric  |    232|        267|     53.51|
#> |Declare       |numeric  |    315|        184|     36.87|
#> |DecDate       |character|     72|        427|     85.57|
#> |Population    |character|    319|        180|     36.07|
#> |PopDate       |numeric  |    316|        183|     36.67|
#> |PopAm         |numeric  |    339|        160|     32.06|
#> |PopulationHigh|numeric  |    139|        360|     72.14|
#> |PopulationLow |numeric  |    108|        391|     78.36|
#> |StartType_Am  |numeric  |    339|        160|     32.06|
#> |StartSettle   |numeric  |    320|        179|     35.87|
#> |End_Am        |numeric  |    499|          0|         0|
#> |EndType_Am    |numeric  |    294|        205|     41.08|
#> |EndSettle     |numeric  |    282|        217|     43.49|
#> |Sovereignty_Am|numeric  |    499|          0|         0|
#> |EuroDip       |numeric  |    331|        168|     33.67|
#> |Borders       |numeric  |    332|        167|     33.47|
#> |Borders_Am    |numeric  |    342|        157|     31.46|
#> |Capital       |character|    284|        215|     43.09|
#> |VioStart      |numeric  |    318|        181|     36.27|
#> |VioStart_Am   |numeric  |    327|        172|     34.47|
#> |VioEnd        |numeric  |    292|        207|     41.48|
#> |VioEnd_Am     |numeric  |    297|        202|     40.48|
#> ---------------------------------------------------------
#>
#>
#> $GW
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss %  |
#> -------------------------------------------------------
#> |stateID     |character|    216|          0|         0|
#> |StateName   |character|    216|          0|         0|
#> |Begin       |mdate    |    216|          0|         0|
#> |End         |mdate    |    216|          0|         0|
#> |StateNameAlt|character|     18|        198|     91.67|
#> |cowID       |character|    216|          0|         0|
#> |cowNR       |character|    216|          0|         0|
#> -------------------------------------------------------
#>
#>
#> $GGO
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss %  |
#> -------------------------------------------------------
#> |stateID     |character|    409|          0|         0|
#> |StateName   |character|    409|          0|         0|
#> |Capital     |character|    409|          0|         0|
#> |Begin       |mdate    |    409|          0|         0|
#> |End         |mdate    |    409|          0|         0|
#> |Latitude    |numeric  |    409|          0|         0|
#> |Longitude   |numeric  |    409|          0|         0|
#> |Region      |character|    409|          0|         0|
#> |StateNameAlt|character|     61|        348|     85.09|
#> |CapitalAlt  |character|      7|        402|     98.29|
#> |Coder       |character|    409|          0|         0|
#> |Comments    |character|     72|        337|      82.4|
#> |Source      |character|    136|        273|     66.75|
#> -------------------------------------------------------

Mapping

GGO	GW	ISD

stateID

Begin	Start	Start

End	Finish	Finish

StateName	Name of State	State.Name

cowID	Cow ID	COW.ID

cowNR	Cow NR.	COW.Nr